requirements.txt
# Needed for production runtime
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
alembic
pydantic[email]
python-dotenv
requests


